Rename phoenixExport legacy file, disable ESLint extension rule
